requirements.txt
feedparser
pandas
python-dateutil
requests
yagmail
python-telegram-bot==13.15
PyYAML
markdown
